@ Ayhun: Changes to our SQL Interface
@ Ayhun: Some additional functionalities become desirable as I move
along. I suspect this is not the last (not even close!) time I'm going
to be requesting an update to the SQL interface, but if you can
implement these changes, I'll love you forever.  <3

((Note: Please pull the latest commit, as I moved your library to a new
file and stuff.))